## Summary

Split MinimalAsyncEP dispatch and combine into graph-visible launch and wait
operations. Launches perform metadata preparation and start symmetric-memory
row copies; waits only join the current stream to completion and return aliases
of the pending outputs.

Keepalive inputs describe storage read asynchronously by the communication
stream. Routing metadata consumed by later operations remains ordinary graph
dataflow rather than an artificial wait dependency.

## Why

A monolithic custom operator hides the useful scheduling boundary and forces
communication to complete before independent expert work can run. Thin aliasing
waits expose the boundary while remaining compatible with autograd,
functionalization, full Inductor, and CUDA graph wrapping.

## Test Plan

- Trace the registered production APIs with fake tensors and verify launch/wait
  dependency edges.
- Validate wait aliasing through structured operator schema metadata.
- Compile wait forward and backward paths with full Inductor.
- Compare two-rank outputs and gradients with a local reference.
- Trace the production APIs through the CUDA graph pass.

[ghstack-poisoned]

+++ torchtitan/distributed/minimal_async_ep/kernels.py
@@ -171,7 +171,7 @@ def _reduce_topk_slots_kernel(
         ``scores=[0.1, 0.2, 0.3, 0.4]``, the output is
         ``[[10], [15]]``.
     """
-    token = tl.program_id(0)
+    token = tl.program_id(0).to(tl.int64)
     col = tl.program_id(1) * BLOCK_N + tl.arange(0, BLOCK_N)
     col_mask = col < NUM_COLS
     acc = tl.zeros((BLOCK_N,), tl.float32)
@@ -223,7 +223,7 @@ def _expand_topk_grad_kernel(
         ``scores=[0.1, 0.2, 0.3, 0.4]``, the output is
         ``grad_routed=[[60], [10], [80], [20]]``.
     """
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)
     col = tl.program_id(1) * BLOCK_N + tl.arange(0, BLOCK_N)
     col_mask = col < NUM_COLS
     flat_index = tl.load(flat_indices + row)
@@ -269,7 +269,7 @@ def _topk_scores_grad_kernel(
         ``flat_indices=[2, 0, 3, 1]``, the slot-ordered output is
         ``grad_scores=[2000, 4000, 2000, 6000]``.
     """
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)
     col = tl.arange(0, BLOCK_N)
     col_mask = col < NUM_COLS
     flat_index = tl.load(flat_indices + row)
